Adding JDK 1.8 pom for use on old Pi3b+ test board
 * refactored non-blocking port services into single class
 * extending Args with a base Settings() class to enable setting of environment variables for greating configuration felxability
 * improved data handling around missing altitudes on some records.
 * update data model types

 Changes to be committed:
	modified:   README.md
	new file:   build.sh
	modified:   pom.xml
	new file:   pom8.xml
	typechange: sbin/monitor5.sh
	modified:   src/main/java/com/daviddemartini/stratux/miltracker/StratuxMilTracker.java
	deleted:    src/main/java/com/daviddemartini/stratux/miltracker/comms/stratux/InRangeMilContactFeed.java
	deleted:    src/main/java/com/daviddemartini/stratux/miltracker/comms/stratux/MilContactFeed.java
	renamed:    src/main/java/com/daviddemartini/stratux/miltracker/comms/stratux/InRangeFeed.java -> src/main/java/com/daviddemartini/stratux/miltracker/comms/stratux/MutliThreadChannelSocket.java
	deleted:    src/main/java/com/daviddemartini/stratux/miltracker/comms/stratux/NewContactFeed.java
	modified:   src/main/java/com/daviddemartini/stratux/miltracker/datamodel/AircraftSBS.java
	modified:   src/main/java/com/daviddemartini/stratux/miltracker/util/Args.java
	new file:   src/main/java/com/daviddemartini/stratux/miltracker/util/Settings.java
	modified:   src/test/java/com/daviddemartini/stratux/miltracker/ArgsTest.java
	new file:   src/test/java/com/daviddemartini/stratux/miltracker/SettingsTest.java

+++ src/main/java/com/daviddemartini/stratux/miltracker/datamodel/AircraftSBS.java
@@ -47,11 +47,11 @@ public class AircraftSBS {
     // the following fields are not expected part of every message, and contain aircraft details
 
     private String callsign;
-    private int altitude;
+    private Integer altitude;
     private Float speedGround;
     private float track;
-    private double latitude;
-    private double longitude;
+    private Double latitude;
+    private Double longitude;
     private int verticalRate;
     private int squawkInt;
     private Boolean squawkChanged;
@@ -304,7 +304,7 @@ public class AircraftSBS {
         return callsign;
     }
 
-    public long getAltitude() {
+    public Integer getAltitude() {
         return altitude;
     }
 
@@ -316,11 +316,11 @@ public class AircraftSBS {
         return track;
     }
 
-    public double getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public double getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
